requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.1.0
audioread==3.0.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
coverage==7.6.12
crepe==0.0.16
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
Deprecated==1.2.18
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hmmlearn==0.3.3
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
itsdangerous==2.2.0
japanize-matplotlib==1.1.3
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
librosa==0.9.2
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mcp==1.6.0
mdurl==0.1.2
mido==1.3.3
mir_eval==0.8.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==1.4.3
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
pretty_midi==0.2.9
prettytable==3.15.1
protobuf==5.29.3
py==1.11.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pytest==7.1.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-rtmidi==1.5.8
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.32.3
resampy==0.4.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.12.2
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
sse-starlette==2.2.1
starlette==0.46.1
stockwell==1.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
uv==0.6.12
uvicorn==0.34.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
